Refactoring `check_pulse_queue_exists()` and add job exception listener
