+++ ejenti/pulse_modules/tasksTrigger.py
@@ -10,6 +10,7 @@ import urlparse
 from datetime import datetime
 from logging.handlers import TimedRotatingFileHandler
 from apscheduler.schedulers.background import BackgroundScheduler
+from apscheduler.events import EVENT_JOB_EXECUTED, EVENT_JOB_ERROR
 
 from hasal_consumer import HasalConsumer
 from hasalPulsePublisher import HasalPulsePublisher
@@ -226,34 +227,6 @@ class TasksTrigger(object):
         return True
 
     @staticmethod
-    def check_pulse_queue_exists(username, password, topic):
-        """
-        Checking does the Queue of Topic exist, if not, then create Queue for Topic on Pulse.
-        Note: If there is no Queues listen on topic, the message will be ignored by Pulse.
-        @param username: Pulse username.
-        @param password: Pulse password.
-        @param topic: Topic.
-        @return: True if queue exists or be created successfully. False if not.
-        """
-        def got_msg(body, message):
-            # does not ack, so broker will keep this message
-            logging.debug('do nothing here')
-        c = HasalConsumer(user=username, password=password, applabel=topic)
-        c.configure(topic=topic, callback=got_msg)
-        queue_exist = c.queue_exists()
-        if not queue_exist:
-            logging.warn('Queue not exists, try to declare queue on Topic [{}]'.format(topic))
-            try:
-                # declare the Queue by building consumer
-                c._build_consumer()  # NOQA
-            except Exception as e:
-                logging.error(e)
-                return False
-        queue_exists = c.queue_exists()
-        logging.debug('Pulse Queue on Topic [{}] exists ... {}'.format(topic, queue_exists))
-        return True
-
-    @staticmethod
     def job_pushing_meta_task(username, password, command_config, job_name, topic, amount, platform_build, cmd_name, overwrite_cmd_config=None):
         """
         [JOB]
@@ -271,9 +244,9 @@ class TasksTrigger(object):
         changed = TasksTrigger.check_latest_info_json_md5_changed(job_name=job_name, platform=platform_build)
         if changed:
             # check queue
-            queue_exists = TasksTrigger.check_pulse_queue_exists(username=username,
-                                                                 password=password,
-                                                                 topic=topic)
+            queue_exists = HasalPulsePublisher.check_pulse_queue_exists(username=username,
+                                                                        password=password,
+                                                                        topic=topic)
             if not queue_exists:
                 logging.error('There is not Queue for Topic [{topic}]. Message might be ignored.'.format(topic=topic))
 
@@ -357,10 +330,21 @@ class TasksTrigger(object):
 
         c.listen()
 
+    def _job_exception_listener(self, event):
+        if event.exception:
+            logging.error("Job [%s] crashed [%s]" % (event.job_id, event.exception))
+            logging.error(event.traceback)
+
+    def _add_event_listener(self):
+        self.scheduler.add_listener(self._job_exception_listener, EVENT_JOB_EXECUTED | EVENT_JOB_ERROR)
+
     def run(self):
         """
         Adding jobs into scheduler.
         """
+        # add event listener
+        self._add_event_listener()
+
         # create "mgt" channel listener
         logging.info('Adding Rotating Logger for listen Agent information ...')
         MGT_ID = 'trigger_mgt_listener'
